+++ lsiutil/install.sh
@@ -6,24 +6,13 @@
 # See /LICENSE for more information.
 #
 
-install_addon() {
+if [ "${1}" = "late" ]; then
   echo "Installing addon lsiutil - ${1}"
   mkdir -p "/tmpRoot/usr/arc/addons/"
   cp -pf "${0}" "/tmpRoot/usr/arc/addons/"
   
   cp -pf /usr/sbin/lsiutil /tmpRoot/usr/sbin/lsiutil
-}
-
-uninstall_addon() {
+elif [ "${1}" = "uninstall" ]; then
   echo "Uninstalling addon lsiutil - ${1}"
   rm -f "/tmpRoot/usr/sbin/lsiutil"
-}
-
-case "${1}" in
-  late)
-    install_addon "${1}"
-    ;;
-  uninstall)
-    uninstall_addon "${1}"
-    ;;
-esac
+fi
